requirements.txt
Django==2.1
django-haystack==2.8.1
pkg-resources==0.0.0
pytz==2019.3
urllib3==1.22
Whoosh==2.7.4
